seeding script seperated from database index.js, and server route for /houses created to get house data
