+++ database/index.js
@@ -1,5 +1,3 @@
-const data = require('./testData.js')
-
 const mongoose = require('mongoose');
 mongoose.connect('mongodb://localhost/checkout');
 
@@ -23,21 +21,4 @@ let repoSchema = mongoose.Schema({
 
 let House = mongoose.model('house', repoSchema);
 
-let insertTestData = function() {
-    House.insertMany(data, (err) => {
-        if (err) {
-            console.log(error)
-        } else {
-            console.log("data was inserted")
-        }
-        mongoose.connection.close()
-    });
-}
-
-House.remove({}, (err, result) => {
-    if (err) {
-        console.log(err);
-    } else {
-        insertTestData()
-    }
-});
+module.exports.House = House
