+++ database/seedingScript.js
@@ -1,0 +1,20 @@
+const db = require('./index.js')
+const data = require('./testData.js')
+
+let insertTestData = function() {
+    db.House.insertMany(data, (err) => {
+        if (err) {
+            console.log(error)
+        } else {
+            console.log("data was inserted")
+        }
+    });
+}
+
+db.House.remove({}, (err, result) => {
+    if (err) {
+        console.log(err);
+    } else {
+        insertTestData()
+    }
+});
